api/api/views.py
from rest_framework.permissions import AllowAny
from rest_framework import generics
from rest_framework import viewsets
from .serializers import TaskSerializer, UserSerializer
from .models import Task


class CreateUserView(generics.CreateAPIView):
    """POST"""
    serializer_class = UserSerializer
    permission_classes = (AllowAny,)


class MyProfileView(generics.RetrieveUpdateAPIView):
    """GET"""
    serializer_class = UserSerializer

    def get_object(self):
        return self.request.user


class TaskViewSet(viewsets.ModelViewSet):
    """GET POST PUT DELETE"""
    queryset = Task.objects.all()
    serializer_class = TaskSerializer
